first working version

+++ pyHomeVM/__main__.py
@@ -1,3 +1,4 @@
+"""__Main__."""
 import sys
 import os
 import logging
@@ -10,13 +11,14 @@ from settings.settings import load_config, load_core, load_remote, load_email
 from settings.settings import load_html, load_sms
 from core import read_structure, readStructureFromFile, updateStructure
 from core import clean_video_db, syncDirTree, transferLongVersions
-from core import executeToDoFile, build_html_report
+from core import executeToDoFile, build_html_report, umount
 from core import check_and_correct_videos_errors, clean_remote
 from core import get_new_file_ids_from_structure, mount, check_mkv_videos
 from notifications import sendSmsNotification, sendMailReport, sendMailLog
 
 
 def get_args():
+    """Get args."""
     parser = argparse.ArgumentParser(description='pyHomeVM')
     parser.add_argument('-c', '--config_file_path',
                         action='store',
@@ -35,11 +37,15 @@ def get_args():
     parser.add_argument('-b', '--backup',
                         help='Enables backup of first videos',
                         action='store_true')
+    parser.add_argument('-stats',
+                        help='Gets you statistics about your videos',
+                        action='store_true')
     args = parser.parse_args()
     return args
 
 
 def load_logger():
+    """Load logger."""
     logger = logging.getLogger(__name__)
     logger.setLevel(logging.DEBUG)
     handler = logging.FileHandler(CONSTANTS['log_file_path'])
@@ -50,10 +56,21 @@ def load_logger():
 
 
 def main(argv=None):
+    """Run main."""
     start_time = datetime.now()
     args = get_args()  # Get args
     logger = load_logger()  # Set logger
+    logger.info('PROGRAM STARTED')
+    pid = str(os.getpid())
+    pidfile = "/tmp/pyHomeVM.pid"
     config = load_config(args.config_file_path)  # load config file
+    if os.path.isfile(pidfile):
+        logger.info('Program already running')
+        html = load_html(config)
+        email = load_email(config)
+        sendMailLog(CONSTANTS['log_file_path'], email, html)
+        sys.exit()
+    file(pidfile, 'w').write(pid)
     (ffmpeg, local) = load_core(config)  # load core configs
     remote = load_remote(config)
     html = load_html(config)
@@ -88,8 +105,7 @@ def main(argv=None):
         remote,
         video_db)
     sms_sent_file = os.path.join(CONSTANTS['script_root_dir'], 'sms_sent')
-    #if(mount(remote)):
-    if(True):
+    if(mount(remote)):
         logger.info('Mount succesfull')
         syncDirTree(local, remote)
         transferLongVersions(local, remote, video_db)
@@ -99,10 +115,11 @@ def main(argv=None):
             os.remove(sms_sent_file)
             logger.info('sms_sent file has been deleted')
         clean_remote(remote)
+        umount(remote)
     else:
         logger.info('Mount unssuccesfull')
-        if(not os.path.exists(sms_sent_file)):
-            #sendSmsNotification(sms)
+        if(not os.path.exists(sms_sent_file) and args.sms):
+            sendSmsNotification(sms)
             logger.info('Sms sent')
             with open(sms_sent_file, 'w') as sms_not:
                 msg = 'SMS has been sent {}'.format(CONSTANTS['TODAY'])
@@ -117,12 +134,13 @@ def main(argv=None):
         sendMailReport(html_report, email)
         logger.info('Mail report sent')
     if(args.log):
-        sendMailLog(CONSTANTS['log_file_path'], email)
+        sendMailLog(CONSTANTS['log_file_path'], email, html)
         logger.info('log file sent')
     clean_video_db(video_db)
     check_mkv_videos(local, video_db)
     logger.info('DB cleaned')
     video_db.close()
     logger.info('Script ran in {}'.format(datetime.now() - start_time))
+    os.unlink(pidfile)
 if __name__ == "__main__":
     sys.exit(main())
